services/ingest_chunker/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-cloud-documentai==2.28.0
google-cloud-bigquery==3.25.0
